feat(gazelle): Add resolve integration tests
